+++ src/workflow/oph_workflow_structs.c
@@ -1,6 +1,6 @@
 /*
     Ophidia Terminal
-    Copyright (C) 2012-2020 CMCC Foundation
+    Copyright (C) 2012-2021 CMCC Foundation
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -106,6 +106,10 @@ int oph_workflow_free(oph_workflow * workflow)
 	if (workflow->host_partition) {
 		free(workflow->host_partition);
 		workflow->host_partition = NULL;
+	}
+	if (workflow->project) {
+		free(workflow->project);
+		workflow->project = NULL;
 	}
 	free(workflow);
 	workflow = NULL;